rename Fade -> BlockFade
