+++ system/animations/FadeManager.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import * as React from 'react';
-import { FadeConfig } from './Fade';
+import { FadeConfig } from './BlockFade';
 
 export type FadeParamFn = (index: number) => number;
 export type FadeParamInterval = { initial?: number, interval?: number };
